TCF Agent: fixed issues with building and running the agent on VxWorks: 1. accept() crashes if socket address buffer is bigger thew SOCK_MAXADDRLEN. 2. missing #include <framework/myalloc.h>. 3. conflicting declaration of uintptr_t.

+++ framework/mdep.h
@@ -300,7 +300,7 @@ extern char * canonicalize_file_name(const char * path);
 #define closesocket close
 
 #if _WRS_VXWORKS_MAJOR < 6 || _WRS_VXWORKS_MAJOR == 6 && _WRS_VXWORKS_MINOR < 9
-typedef unsigned long uintptr_t;
+typedef unsigned int uintptr_t;
 #define send(s, buf, len, flags) (send)(s, (char *)(buf), len, flags)
 #endif
